Migration of check/varfont/bold_wght_coord
From the Open Type profile to the Google Fonts profile
com.google.fonts/check/varfont/bold_wght_coord

(issue #4436)
