+++ Lib/fontbakery/profiles/opentype.py
@@ -46,7 +46,6 @@ OPENTYPE_PROFILE_CHECKS = [
     "com.google.fonts/check/varfont/regular_slnt_coord",
     "com.google.fonts/check/varfont/regular_ital_coord",
     "com.google.fonts/check/varfont/regular_opsz_coord",
-    "com.google.fonts/check/varfont/bold_wght_coord",
     "com.google.fonts/check/varfont/slnt_range",
     "com.typenetwork/check/varfont/ital_range",
     "com.google.fonts/check/varfont/wght_valid_range",
